feat: move remotelink command to a subcommand of link (#446)
