+++ pkg/jira/issue.go
@@ -380,3 +380,42 @@ func ifaceToADF(v interface{}) *adf.ADF {
 
 	return doc
 }
+
+type remotelinkRequest struct {
+	RemoteObject struct {
+		URL   string `json:"url"`
+		Title string `json:"title"`
+	} `json:"object"`
+}
+
+// RemoteLinkIssue adds a remote link to an issue using POST /issue/{issueId}/remotelink endpoint.
+func (c *Client) RemoteLinkIssue(issueID, title, url string) error {
+	body, err := json.Marshal(remotelinkRequest{
+		RemoteObject: struct {
+			URL   string `json:"url"`
+			Title string `json:"title"`
+		}{Title: title, URL: url},
+	})
+	if err != nil {
+		return err
+	}
+
+	path := fmt.Sprintf("/issue/%s/remotelink", issueID)
+
+	res, err := c.PostV2(context.Background(), path, body, Header{
+		"Accept":       "application/json",
+		"Content-Type": "application/json",
+	})
+	if err != nil {
+		return err
+	}
+	if res == nil {
+		return ErrEmptyResponse
+	}
+	defer func() { _ = res.Body.Close() }()
+
+	if res.StatusCode != http.StatusCreated {
+		return formatUnexpectedResponse(res)
+	}
+	return nil
+}
